requirements.txt
# Google Cloud SDK
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-cloud-secret-manager==2.24.0

# Web Framework
fastapi==0.116.1
uvicorn==0.35.0

# HTTP Client
httpx==0.28.1
requests==2.32.4

# Environment Management
python-dotenv==1.1.1

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==4.1.0

# Utilities
pydantic==2.11.7
pydantic-settings==2.10.1
